// src/features/vip/tier/data/schema.ts
export type VipTier = {
    id: number
    name: string
    mb_b2b_point: number
    mb_b2b_reward: number
    b2c_point: number
    b2c_reward: number
    b2b_point: number
    b2b_reward: number
    sort_order: number
    status: number
    note?: string
    created_at?: string
    updated_at?: string
}